extract ouput generation function from index
